gpt-image-1/requirements.txt
openai>=1.77.0
python-dotenv>=1.1.0
Pillow>=11.2.1

